move client commands into one place under specific classes
(cherry picked from commit 1888149ec1e1a0729aef6d00cf891bff94bbe37a)
