+++ django_valkey/client/default.py
@@ -1,6 +1,7 @@
 from valkey import Valkey
-from django_valkey.base_client import BaseClient
+
+from django_valkey.base_client import BaseClient, ClientCommands
 
 
-class DefaultClient(BaseClient[Valkey]):
+class DefaultClient(BaseClient[Valkey], ClientCommands[Valkey]):
     CONNECTION_FACTORY_PATH = "django_valkey.pool.ConnectionFactory"
